Split associations setup into own file
